Support for protobuf and moving zipkin forwarded to new package
